+++ src/rqt_gauges_2/steering_wheel_gauge.py
@@ -37,8 +37,8 @@ class SteeringWheelGauge(QWidget):
         self.widget_diameter = min(self.width, self.height)
         self.resize(self.width, self.height)
 
-        self.scale_line_outer_start = (self.widget_diameter)/2
-        self.scale_line_length = (self.widget_diameter / 2) - (self.widget_diameter / 20)
+        self.scale_line_outer_start = int(self.widget_diameter / 2)
+        self.scale_line_length = int((self.widget_diameter / 2) - (self.widget_diameter / 20))
 
         self.text_radius_factor = 0.75
         self.text_radius = self.widget_diameter/2 * self.text_radius_factor
@@ -55,7 +55,7 @@ class SteeringWheelGauge(QWidget):
 
         painter.rotate(start_angle_value - self.angle_offset)
         steps_size = (float(angle_size) / float(scala_count))
-        
+
         for _ in range(scala_count+1):
             painter.drawLine(self.scale_line_length, 0, self.scale_line_outer_start, 0)
             painter.rotate(steps_size)
@@ -81,8 +81,8 @@ class SteeringWheelGauge(QWidget):
             h = fm.height()
             painter.setFont(QFont(self.font_family, self.scale_font_size))
             angle = angle_distance * i + float(scale_angle_start_value - self.angle_offset)
-            x = self.text_radius * math.cos(math.radians(angle))
-            y = self.text_radius * math.sin(math.radians(angle))
+            x = int(self.text_radius * math.cos(math.radians(angle)))
+            y = int(self.text_radius * math.sin(math.radians(angle)))
 
             text = [x - int(w/2), y - int(h/2), int(w), int(h), Qt.AlignCenter, text]
             painter.drawText(text[0], text[1], text[2], text[3], text[4], text[5])
@@ -90,7 +90,7 @@ class SteeringWheelGauge(QWidget):
 
     def updateValue(self, value: float):
         # Updates the value that the gauge is indicating.
-        # Args: 
+        # Args:
         #   value: Value to update the gauge with.
         value = max(value, self.minValue)
         value = min(value, self.maxValue)
@@ -109,8 +109,8 @@ class SteeringWheelGauge(QWidget):
         # Size variables
         width = self.width - self.progress_width
         height = self.height - self.progress_width
-        margin = self.progress_width / 2
-        value = self.value * self.circle_max_angle / self.max_value
+        margin = int(self.progress_width / 2)
+        value = int(self.value * self.circle_max_angle / self.max_value)
 
         # Draw Circle
         self.draw_background_circle()
@@ -147,4 +147,3 @@ class SteeringWheelGauge(QWidget):
 
         self.draw_big_scaled_marker(131, 278, 5)
         self.create_scale_marker_values_text(131, 278, 5)
-
